new-packagesystem: integrate with bootupd, AND USE CARGO FMT
