+++ src/main.rs
@@ -45,7 +45,6 @@ mod model;
 mod model_legacy;
 mod ostreeutil;
 mod packagesystem;
-mod packagesystem_fileonly;
 mod sha512string;
 mod util;
